backend/requirements.txt
requests==2.26.0
django==3.2.18
djangorestframework==3.12.4
django-filter==21.1
djoser==2.1.0
flake8==5.0.4
Pillow==9.4.0
psycopg2-binary==2.9.5
sorl-thumbnail==12.9.0